Prevent leaking NXEventHandles
This commit is important, as I suspect that this is causing the GPU crashes and
kernel panics. Yes, leaking handles can and will cause your kernel to panic for
some reason, but only on High Sierra (10.13). No, I have no idea why this
happens. Probably some strange bug in IOKit that I know nothing about.

Hopefully this fixes it.

Hammerspoon code was used as reference:
https://github.com/Hammerspoon/hammerspoon/blob/master/extensions/mouse/internal.m#L67